clang tidy

+++ src/fheroes2/gui/interface_list.h
@@ -467,7 +467,7 @@ namespace Interface
             if ( !_lockClick && le.MouseLongPressLeft( rtAreaItems ) && content->size() > 1 ) {
                 const fheroes2::Point & mousePos = le.getMouseCursorPos();
                 const int id = ( mousePos.y - rtAreaItems.y ) * maxItems / rtAreaItems.height + _topId;
-                if ( id <= ( content->size() - 1 ) ) {
+                if ( static_cast<size_t>( id ) <= ( content->size() - 1 ) ) {
                     Item & item = ( *content )[static_cast<size_t>( id )]; // id is always >= 0
                     ActionListLongPress( item );
                 }
